+++ packages/token-balance-cache/.mocharc.json
@@ -1,0 +1,7 @@
+{
+  "extension": [".spec.ts"],
+  "package": "./package.json",
+  "timeout": 60000,
+  "require": [ "tsconfig-paths/register", "source-map-support/register" ],
+  "exit": true
+}
