+++ packages/token-balance-cache/test/providers/cosmos.spec.ts
@@ -1,0 +1,8 @@
+import { assert } from 'chai';
+import { default as CosmosBalanceProvider } from '../../src/providers/cosmos'
+
+describe('CosmosBalanceProvider unit tests', () => {
+  it('should TODO', () => {
+    
+  })
+});
