+++ packages/token-balance-cache/test/tbc.spec.ts
@@ -1,0 +1,8 @@
+import { assert } from 'chai';
+import { default as TBC } from '../src/tbc'
+
+describe('TBC unit tests', () => {
+  it('should TODO', () => {
+    
+  })
+});
